redo changes navBar and moving css files
