+++ src/styles/RankingCard.css
